\name{laeken-package}
\alias{laeken-package}
\alias{laeken}
\docType{package}
\title{
Laeken indicators for measuring social cohesion
}
\description{
Calculate Laeken indicators for measuring social cohesion and apply Pareto tail modeling to empirical income distributions (including graphical tools).
}
\details{
\tabular{ll}{
Package: \tab laeken\cr
Type: \tab Package\cr
Version: \tab 0.1.2\cr
Date: \tab 2010-08-09\cr
Depends: \tab boot, MASS\cr
License: \tab GPL (>= 2)\cr
}

Index:
\preformatted{
arpr                    At-risk-of-poverty rate
arpt                    At-risk-of-poverty threshold
bootVar                 Bootstrap variance and confidence intervals of
                        Laeken indicators
calibWeights            Calibrate sample weights
eqInc                   Equivalized disposable income
eqSS                    Equivalized household size
eusilc                  Synthetic EU-SILC survey data
fitPareto               Fit income distribution models with the Pareto
                        distribution
gini                    Gini coefficient
incMedian               Weighted median income
incQuintile             Weighted income quintile
is.indicator            Utility functions for Laeken indicators
meanExcessPlot          Mean excess plot
minAMSE                 Weighted asymptotic mean squared error (AMSE)
                        estimator
paretoQPlot             Pareto quantile plot
qsr                     Quintile share ratio
rmpg                    Relative median at-risk-of-poverty gap
thetaHill               Hill estimator
thetaISE                Integrated squared error (ISE) estimator
thetaLS                 Least squares (LS) estimator
thetaMoment             Moment estimator
thetaPDC                Partial density component (PDC) estimator
thetaQQ                 QQ-estimator
thetaTM                 Trimmed mean estimator
thetaWML                Weighted maximum likelihood estimator
variance                Variance and confidence intervals of Laeken
                        indicators
weightedMedian          Weighted median
weightedQuantile        Weighted quantiles
}
}
\author{
Andreas Alfons, Josef Holzer and Matthias Templ

Maintainer: Andreas Alfons, \email{alfons@statistik.tuwien.ac.at}
}
\keyword{package}
