\name{laeken-package}
\alias{laeken-package}
\alias{laeken}
\docType{package}
\title{
\packageTitle{laeken}
}
\description{
\packageDescription{laeken}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{laeken}
\packageIndices{laeken}
}
\references{
A. Alfons and M. Templ (2013) Estimation of Social Exclusion Indicators
from Complex Surveys: The \R Package \pkg{laeken}.  \emph{Journal of
Statistical Software}, \bold{54}(15), 1--25.  \doi{10.18637/jss.v054.i15}

A. Alfons, M. Templ, P. Filzmoser (2013) Robust estimation of economic
indicators from survey samples based on Pareto tail modeling. \emph{Journal of
the Royal Statistical Society, Series C}, \bold{62}(2), 271--286.
}
\author{
\packageAuthor{laeken}

Maintainer: \packageMaintainer{laeken}
}
\keyword{package}
