\name{laeken-package}
\alias{laeken-package}
\alias{laeken}
\docType{package}
\title{
Estimation of indicators on social exclusion and poverty
}
\description{
Estimation of indicators on social exclusion and poverty and Pareto tail modeling for empirical income distributions (including graphical tools).
}
\details{
\tabular{ll}{
Package: \tab laeken\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2011-03-08\cr
Depends: \tab boot, MASS\cr
License: \tab GPL (>= 2)\cr
}

Index:
\preformatted{
arpr                    At-risk-of-poverty rate
arpt                    At-risk-of-poverty threshold
bootVar                 Bootstrap variance and confidence intervals of
                        indicators on social exclusion and poverty
calibVars               Construct a matrix of binary variables for
                        calibration
calibWeights            Calibrate sample weights
eqInc                   Equivalized disposable income
eqSS                    Equivalized household size
eusilc                  Synthetic EU-SILC survey data
fitPareto               Fit income distribution models with the Pareto
                        distribution
gini                    Gini coefficient
incMean                 Weighted mean income
incMedian               Weighted median income
incQuintile             Weighted income quintile
is.indicator            Utility functions for indicators on social
                        exclusion and poverty
laeken-package          Estimation of indicators on social exclusion
                        and poverty
meanExcessPlot          Mean excess plot
minAMSE                 Weighted asymptotic mean squared error (AMSE)
                        estimator
paretoQPlot             Pareto quantile plot
paretoScale             Estimate the scale parameter of a Pareto
                        distribution
paretoTail              Pareto tail modeling for income distributions
qsr                     Quintile share ratio
replaceTail             Replace observations under a Pareto model
reweightOut             Reweight outliers in the Pareto model
rmpg                    Relative median at-risk-of-poverty gap
thetaHill               Hill estimator
thetaISE                Integrated squared error (ISE) estimator
thetaLS                 Least squares (LS) estimator
thetaMoment             Moment estimator
thetaPDC                Partial density component (PDC) estimator
thetaQQ                 QQ-estimator
thetaTM                 Trimmed mean estimator
thetaWML                Weighted maximum likelihood estimator
variance                Variance and confidence intervals of indicators
                        on social exclusion and poverty
weightedMean            Weighted mean
weightedMedian          Weighted median
weightedQuantile        Weighted quantiles
}

Further information is available in the following vignettes:
\tabular{ll}{
\code{laeken-pareto} \tab Robust Pareto Tail Modeling for the Estimation of Indicators on Social Exclusion using the R Package laeken (source, pdf)\cr
\code{laeken-standard} \tab Standard Methods for Point Estimation of Indicators on Social Exclusion and Poverty using the R Package laeken (source, pdf)\cr
\code{laeken-variance} \tab Variance Estimation of Indicators on Social Exclusion and Poverty using the R Package laeken (source, pdf)\cr
}
}
\author{
Andreas Alfons, Josef Holzer and Matthias Templ

Maintainer: Andreas Alfons, \email{alfons@statistik.tuwien.ac.at}
}
\keyword{package}
